move utils to inference/retrieval_utils.py
